+++ src/LiteGraph.Sdk/Edge.cs
@@ -80,7 +80,7 @@
         /// <summary>
         /// Labels.
         /// </summary>
-        public List<LabelMetadata> Labels { get; set; } = null;
+        public List<string> Labels { get; set; } = null;
 
         /// <summary>
         /// Tags.